requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
textblob==0.17.1
python-dotenv==1.0.0
requests==2.31.0
Flask-Login==0.6.3
